Separated knight movement to a different file
